Narrow String::CharAt from int32_t to uint16_t.
R=asiva@google.com

Review URL: https://codereview.chromium.org//560113002

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge/dart@40178 260f80e4-7a28-3924-810f-c04153c831b5

+++ runtime/vm/dart_api_impl.cc
@@ -2306,7 +2306,7 @@ DART_EXPORT Dart_Handle Dart_StringToUTF16(Dart_Handle str,
   intptr_t str_len = str_obj.Length();
   intptr_t copy_len = (str_len > *length) ? *length : str_len;
   for (intptr_t i = 0; i < copy_len; i++) {
-    utf16_array[i] = static_cast<uint16_t>(str_obj.CharAt(i));
+    utf16_array[i] = str_obj.CharAt(i);
   }
   *length = copy_len;
   return Api::Success();
@@ -2373,7 +2373,7 @@ DART_EXPORT Dart_Handle Dart_MakeExternalString(Dart_Handle str,
       ASSERT(length >= (copy_len * str_obj.CharSize()));
       uint16_t* utf16_array = reinterpret_cast<uint16_t*>(array);
       for (intptr_t i = 0; i < copy_len; i++) {
-        utf16_array[i] = static_cast<uint16_t>(str_obj.CharAt(i));
+        utf16_array[i] = str_obj.CharAt(i);
       }
       TwoByteString::SetPeer(str_obj, peer, cback);
     }
